requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
python-dotenv==1.0.0
pymongo==4.5.0
langchain==0.0.312
openai==0.28.1
chromadb==0.4.14
pypdf==3.16.4
tiktoken==0.5.1
